Plugin loader merged with file accessor implementation.
