+++ earth_enterprise/src/common/AbstractFileAccessor.h
