+++ earth_enterprise/src/common/FileAccessor.h
@@ -12,11 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#ifndef COMMON_FILEACCESSOR_H
+#define COMMON_FILEACCESSOR_H
+
 #include <string>
 #include <cstring>
 #include <memory>
 #include <khFileUtils.h>
 #include "khStringUtils.h"
+#include "AbstractFileAccessor.h"
 
 class AbstractFileAccessor {
 public:
@@ -36,26 +40,4 @@ public:
   virtual void fprintf(const char *format, ...) = 0;
 };
 
-class POSIXFileAccessor: public AbstractFileAccessor {
-  int fileDescriptor;
-public:
-  POSIXFileAccessor(int i = -1) : fileDescriptor{ i }{}
-  ~POSIXFileAccessor(void) {
-    if (this->isValid()) {
-      this->Close();
-    }
-  }
-
-  bool isValid() override { return fileDescriptor != -1; }
-  void invalidate() override { fileDescriptor = -1; }
-  int getFD() override { return fileDescriptor; }
-  void Open(const std::string &fname, const char *mode = nullptr, int flags = 00, mode_t createMask = 0666) override;
-  int FsyncAndClose() override;
-  int Close() override;
-  bool PreadAll(void* buffer, size_t size, off64_t offset) override;
-  bool PwriteAll(const void* buffer, size_t size, off64_t offset) override;
-  bool ReadStringFromFile(const std::string &filename, std::string &str, std::uint64_t limit = 0) override;
-  bool Exists(const std::string &filename) override;
-  bool GetLinesFromFile(std::vector<std::string> &lines, const std::string &filename) override;
-  void fprintf(const char *format, ...) override;
-};
+#endif //COMMON_FILEACCESSOR_H
